c_items_test: fixed 64-bit MSVC warnings

+++ testing/unit-tests/c/c_items_test/gen1_items_tests.c
@@ -223,7 +223,7 @@ static void gen1_item_bag_test(
         error = pkmn_item_bag_add(
                     &bag,
                     ITEM_NAMES[i],
-                    i+1
+                    (int)(i+1)
                 );
         TEST_ASSERT_EQUAL(PKMN_ERROR_NONE, error);
     }
@@ -243,7 +243,7 @@ static void gen1_item_bag_test(
         error = pkmn_item_bag_remove(
                     &bag,
                     ITEM_NAMES[i],
-                    i+1
+                    (int)(i+1)
                 );
         TEST_ASSERT_EQUAL(PKMN_ERROR_NONE, error);
     }
